requirements.txt
Django==5.2.3
psycopg2-binary==2.9.10
django-mptt==0.17.0
django-debug-toolbar==6.0.0
django-widget-tweaks==1.5.0
# whitenoise==6.9.0
# sentry-sdk==2.34.1
python-dotenv==1.1.1 #settings load_dotenv()
